Ensure user-entered SBP is a float
Ensure that a user-entered systolic blood pressure is parsed as a float rather than a string.

Previously, if the systolic blood pressure was retrieved from the FHIR server, it was stored internally as a float. However, it was a user-entered value, it was stored as a string. This causes issues if you need to perform logic on this value.